Fixes some cast warnings

+++ UI/UIBox.cpp
@@ -22,10 +22,10 @@ void UIBox::resize(int width, int height) {
 
     // Add the center box
     Vec2f boxPosition(0, 0);
-    Vec2f boxDimensions(_dimensions.x, _dimensions.y);
+    Vec2f boxDimensions((float)_dimensions.x, (float)_dimensions.y);
     if(_uiBorder > 0) {
         // Subtract space for the border if there is a border that protrudes inwards
-        boxPosition += Vec2f(bWidth, bHeight);
+        boxPosition += Vec2f((float)bWidth, (float)bHeight);
         boxDimensions -= Vec2f(bWidth * 2.0f, bHeight * 2.0f);
     }
     addRenderable(Renderable::OrthoBox(boxPosition, boxDimensions, false, false, _material));
